requirements.txt
# Core dependencies
streamlit==1.28.0
python-dotenv==1.0.0
google-generativeai==0.3.0

# PDF processing
pdfplumber==0.10.3
PyMuPDF>=1.24.0
PyPDF2==3.0.1

# Translation and AI
deep-translator==1.11.4
sympy==1.12

# PDF generation
reportlab==4.0.4
playwright==1.40.0

# Utilities
tqdm==4.66.0